Fixes error when setting snooze, repeat and vibration.

+++ android/src/main/java/com/emekalites/react/alarm/notification/ANModule.java
@@ -62,14 +62,14 @@ public class ANModule extends ReactContextBaseJavaModule {
         Bundle data = bundle.getBundle("data");
         alarm.setData(bundle2string(data));
 
-        alarm.setInterval(bundle.getInt("repeat_interval", 1));
+        alarm.setInterval((int)bundle.getDouble("repeat_interval", 1.0));
         alarm.setLargeIcon(bundle.getString("large_icon", ""));
         alarm.setLoopSound(bundle.getBoolean("loop_sound", false));
         alarm.setMessage(bundle.getString("message", "My Notification Message"));
         alarm.setPlaySound(bundle.getBoolean("play_sound", true));
         alarm.setScheduleType(bundle.getString("schedule_type", "once"));
         alarm.setSmallIcon(bundle.getString("small_icon", "ic_launcher"));
-        alarm.setSnoozeInterval(bundle.getInt("snooze_interval", 1));
+        alarm.setSnoozeInterval((int)bundle.getDouble("snooze_interval", 1.0));
         alarm.setSoundName(bundle.getString("sound_name", null));
         alarm.setSoundNames(bundle.getString("sound_names", null));
         alarm.setTag(bundle.getString("tag", ""));
@@ -77,7 +77,7 @@ public class ANModule extends ReactContextBaseJavaModule {
         alarm.setTitle(bundle.getString("title", "My Notification Title"));
         alarm.setVibrate(bundle.getBoolean("vibrate", true));
         alarm.setHasButton(bundle.getBoolean("has_button", false));
-        alarm.setVibration(bundle.getInt("vibration", 100));
+        alarm.setVibration((int)bundle.getDouble("vibration", 100.0));
         alarm.setUseBigText(bundle.getBoolean("use_big_text", false));
 
         try {
@@ -144,14 +144,14 @@ public class ANModule extends ReactContextBaseJavaModule {
         Bundle data = bundle.getBundle("data");
         alarm.setData(bundle2string(data));
 
-        alarm.setInterval(bundle.getInt("repeat_interval", 1));
+        alarm.setInterval((int)bundle.getDouble("repeat_interval", 1));
         alarm.setLargeIcon(bundle.getString("large_icon"));
         alarm.setLoopSound(bundle.getBoolean("loop_sound", false));
         alarm.setMessage(bundle.getString("message", "My Notification Message"));
         alarm.setPlaySound(bundle.getBoolean("play_sound", true));
         alarm.setScheduleType(bundle.getString("schedule_type", "once"));
         alarm.setSmallIcon(bundle.getString("small_icon", "ic_launcher"));
-        alarm.setSnoozeInterval(bundle.getInt("snooze_interval", 1));
+        alarm.setSnoozeInterval((int)bundle.getDouble("snooze_interval", 1));
         alarm.setSoundName(bundle.getString("sound_name"));
         alarm.setSoundNames(bundle.getString("sound_names"));
         alarm.setTag(bundle.getString("tag"));
@@ -159,7 +159,7 @@ public class ANModule extends ReactContextBaseJavaModule {
         alarm.setTitle(bundle.getString("title", "My Notification Title"));
         alarm.setVibrate(bundle.getBoolean("loop_sound", true));
         alarm.setHasButton(bundle.getBoolean("has_button", false));
-        alarm.setVibration(bundle.getInt("vibration", 100));
+        alarm.setVibration((int)bundle.getDouble("vibration", 100));
         alarm.setUseBigText(bundle.getBoolean("use_big_text", false));
         
         String datetime = bundle.getString("fire_date");
